Import `hierplane` from it's prior location.
This imports the current version of `hierplane` from it's prior
location.  The only tweak involves some slight revisions to
`static.html`, which now lives in `dev/`.

Longer term we'll likely want to make the mechanism for injecting
sample trees a bit easier to work with. Let me know if this gets
hard to work with quickly and we can improve it.