Preserve the spanType, if provided by end users.
This makes it possible for end users (i.e. me integrating it into
AllenNLP) to provide a `spanType`, which is helpful for rendering
punctuation.

We'll likely need to document this in the README, but I'll leave that
for later.

I also bumped the version, since I'll publish another rev so I can
pull it into AllenNLP.